Move SDXL class to own file
