+++ library/common_gui.py
@@ -635,245 +635,6 @@ def run_cmd_training(**kwargs):
     return run_cmd
 
 
-def gradio_advanced_training(headless=False):
-    def noise_offset_type_change(noise_offset_type):
-        if noise_offset_type == 'Original':
-            return (gr.Group.update(visible=True), gr.Group.update(visible=False))
-        else:
-            return (gr.Group.update(visible=False), gr.Group.update(visible=True))
-    with gr.Row():
-        additional_parameters = gr.Textbox(
-            label='Additional parameters',
-            placeholder='(Optional) Use to provide additional parameters not handled by the GUI. Eg: --some_parameters "value"',
-        )
-    with gr.Row():
-        save_every_n_steps = gr.Number(
-            label='Save every N steps',
-            value=0,
-            precision=0,
-            info='(Optional) The model is saved every specified steps',
-        )
-        save_last_n_steps = gr.Number(
-            label='Save last N steps',
-            value=0,
-            precision=0,
-            info='(Optional) Save only the specified number of models (old models will be deleted)',
-        )
-        save_last_n_steps_state = gr.Number(
-            label='Save last N states',
-            value=0,
-            precision=0,
-            info='(Optional) Save only the specified number of states (old models will be deleted)',
-        )
-    with gr.Row():
-        keep_tokens = gr.Slider(
-            label='Keep n tokens', value='0', minimum=0, maximum=32, step=1
-        )
-        clip_skip = gr.Slider(
-            label='Clip skip', value='1', minimum=1, maximum=12, step=1
-        )
-        max_token_length = gr.Dropdown(
-            label='Max Token Length',
-            choices=[
-                '75',
-                '150',
-                '225',
-            ],
-            value='75',
-        )
-        full_fp16 = gr.Checkbox(
-            label='Full fp16 training (experimental)', value=False
-        )
-    with gr.Row():
-        gradient_checkpointing = gr.Checkbox(
-            label='Gradient checkpointing', value=False
-        )
-        shuffle_caption = gr.Checkbox(label='Shuffle caption', value=False)
-        persistent_data_loader_workers = gr.Checkbox(
-            label='Persistent data loader', value=False
-        )
-        mem_eff_attn = gr.Checkbox(
-            label='Memory efficient attention', value=False
-        )
-    with gr.Row():
-        # This use_8bit_adam element should be removed in a future release as it is no longer used
-        # use_8bit_adam = gr.Checkbox(
-        #     label='Use 8bit adam', value=False, visible=False
-        # )
-        xformers = gr.Checkbox(label='Use xformers', value=True)
-        color_aug = gr.Checkbox(label='Color augmentation', value=False)
-        flip_aug = gr.Checkbox(label='Flip augmentation', value=False)
-        min_snr_gamma = gr.Slider(
-            label='Min SNR gamma', value=0, minimum=0, maximum=20, step=1
-        )
-    with gr.Row():
-        bucket_no_upscale = gr.Checkbox(
-            label="Don't upscale bucket resolution", value=True
-        )
-        bucket_reso_steps = gr.Slider(
-            label='Bucket resolution steps', value=64, minimum=1, maximum=128
-        )
-        random_crop = gr.Checkbox(
-            label='Random crop instead of center crop', value=False
-        )
-    
-    with gr.Row():
-        min_timestep = gr.Slider(
-            label='Min Timestep',
-            value=0,
-            step=1,
-            minimum=0,
-            maximum=1000,
-            info='Values greater than 0 will make the model more img2img focussed'
-        )
-        max_timestep = gr.Slider(
-            label='Max Timestep',
-            value=1000,
-            step=1,
-            minimum=0,
-            maximum=1000,
-            info='Values lower than 1000 will make the model more img2img focussed',
-        )
-    
-    with gr.Row():
-        noise_offset_type = gr.Dropdown(
-            label='Noise offset type',
-            choices=[
-                'Original',
-                'Multires',
-            ],
-            value='Original',
-        )
-        with gr.Row(visible=True) as noise_offset_original:
-            noise_offset = gr.Slider(
-                label='Noise offset',
-                value=0,
-                minimum=0,
-                maximum=1,
-                step=0.01,
-                info='recommended values are 0.05 - 0.15',
-            )
-            adaptive_noise_scale = gr.Slider(
-                label='Adaptive noise scale',
-                value=0,
-                minimum=-1,
-                maximum=1,
-                step=0.001,
-                info='(Experimental, Optional) Since the latent is close to a normal distribution, it may be a good idea to specify a value around 1/10 the noise offset.',
-            )
-        with gr.Row(visible=False) as noise_offset_multires:
-            multires_noise_iterations = gr.Slider(
-                label='Multires noise iterations',
-                value=0,
-                minimum=0,
-                maximum=64,
-                step=1,
-                info='enable multires noise (recommended values are 6-10)',
-            )
-            multires_noise_discount = gr.Slider(
-                label='Multires noise discount',
-                value=0,
-                minimum=0,
-                maximum=1,
-                step=0.01,
-                info='recommended values are 0.8. For LoRAs with small datasets, 0.1-0.3',
-            )
-        noise_offset_type.change(
-            noise_offset_type_change,
-            inputs=[noise_offset_type],
-            outputs=[noise_offset_original, noise_offset_multires]
-        )
-    with gr.Row():
-        caption_dropout_every_n_epochs = gr.Number(
-            label='Dropout caption every n epochs', value=0
-        )
-        caption_dropout_rate = gr.Slider(
-            label='Rate of caption dropout', value=0, minimum=0, maximum=1
-        )
-        vae_batch_size = gr.Slider(
-            label='VAE batch size', minimum=0, maximum=32, value=0, step=1
-        )
-    with gr.Row():
-        save_state = gr.Checkbox(label='Save training state', value=False)
-        resume = gr.Textbox(
-            label='Resume from saved training state',
-            placeholder='path to "last-state" state folder to resume from',
-        )
-        resume_button = gr.Button(
-            '📂', elem_id='open_folder_small', visible=(not headless)
-        )
-        resume_button.click(
-            get_folder_path,
-            outputs=resume,
-            show_progress=False,
-        )
-        max_train_epochs = gr.Textbox(
-            label='Max train epoch',
-            placeholder='(Optional) Override number of epoch',
-        )
-        max_data_loader_n_workers = gr.Textbox(
-            label='Max num workers for DataLoader',
-            placeholder='(Optional) Override number of epoch. Default: 8',
-            value='0',
-        )
-    with gr.Row():
-        wandb_api_key = gr.Textbox(
-            label='WANDB API Key',
-            value='',
-            placeholder='(Optional)',
-            info='Users can obtain and/or generate an api key in the their user settings on the website: https://wandb.ai/login',
-        )
-        use_wandb = gr.Checkbox(
-            label='WANDB Logging',
-            value=False,
-            info='If unchecked, tensorboard will be used as the default for logging.',
-        )
-        scale_v_pred_loss_like_noise_pred = gr.Checkbox(
-            label='Scale v prediction loss',
-            value=False,
-            info='Only for SD v2 models. By scaling the loss according to the time step, the weights of global noise prediction and local noise prediction become the same, and the improvement of details may be expected.',
-        )
-    return (
-        # use_8bit_adam,
-        xformers,
-        full_fp16,
-        gradient_checkpointing,
-        shuffle_caption,
-        color_aug,
-        flip_aug,
-        clip_skip,
-        mem_eff_attn,
-        save_state,
-        resume,
-        max_token_length,
-        max_train_epochs,
-        max_data_loader_n_workers,
-        keep_tokens,
-        persistent_data_loader_workers,
-        bucket_no_upscale,
-        random_crop,
-        bucket_reso_steps,
-        caption_dropout_every_n_epochs,
-        caption_dropout_rate,
-        noise_offset_type,
-        noise_offset,
-        adaptive_noise_scale,
-        multires_noise_iterations,
-        multires_noise_discount,
-        additional_parameters,
-        vae_batch_size,
-        min_snr_gamma,
-        save_every_n_steps,
-        save_last_n_steps,
-        save_last_n_steps_state,
-        use_wandb,
-        wandb_api_key,
-        scale_v_pred_loss_like_noise_pred,
-        min_timestep,
-        max_timestep,
-    )
-
-
 def run_cmd_advanced_training(**kwargs):
     run_cmd = ''
     
@@ -1060,22 +821,3 @@ def verify_image_folder_pattern(folder_path):
     log.info(f'Valid image folder names found in: {folder_path}')
     return true_response
 
-### SDXL Parameters class
-class SDXLParameters:
-    def __init__(self, sdxl_checkbox):
-        self.sdxl_checkbox = sdxl_checkbox
-
-        with gr.Accordion(visible=False, open=True, label='SDXL Specific Parameters') as self.sdxl_row:
-            with gr.Row():
-                self.sdxl_cache_text_encoder_outputs = gr.Checkbox(
-                    label='Cache text encoder outputs',
-                    info='Cache the outputs of the text encoders. This option is useful to reduce the GPU memory usage. This option cannot be used with options for shuffling or dropping the captions.',
-                    value=False,
-                )
-                self.sdxl_no_half_vae = gr.Checkbox(
-                    label='No half VAE',
-                    info='Disable the half-precision (mixed-precision) VAE. VAE for SDXL seems to produce NaNs in some cases. This option is useful to avoid the NaNs.',
-                    value=False
-                )
-
-        self.sdxl_checkbox.change(lambda sdxl_checkbox: gr.Accordion.update(visible=sdxl_checkbox), inputs=[self.sdxl_checkbox], outputs=[self.sdxl_row])
